Aider beta 6 (#40)
* Added working

* Added working

* Added working

* Added stuff

* Added stuff

---------
